requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
